+++ src/vs/workbench/contrib/runtimeNotebookKernel/browser/runtimeNotebookKernelService.ts
@@ -3,40 +3,24 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AsyncIterableObject, DeferredPromise, Sequencer } from '../../../../base/common/async.js';
-import { decodeBase64, VSBuffer } from '../../../../base/common/buffer.js';
-import { CancellationToken } from '../../../../base/common/cancellation.js';
-import { Emitter } from '../../../../base/common/event.js';
 import { Disposable } from '../../../../base/common/lifecycle.js';
-import { ResourceMap } from '../../../../base/common/map.js';
-import { URI } from '../../../../base/common/uri.js';
-import { generateUuid } from '../../../../base/common/uuid.js';
 import { IConfigurationService } from '../../../../platform/configuration/common/configuration.js';
-import { ExtensionIdentifier } from '../../../../platform/extensions/common/extensions.js';
 import { InstantiationType, registerSingleton } from '../../../../platform/instantiation/common/extensions.js';
 import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
 import { ILogService } from '../../../../platform/log/common/log.js';
-import { IQuickInputService } from '../../../../platform/quickinput/common/quickInput.js';
-import { ILanguageRuntimeMessageError, ILanguageRuntimeMessageInput, ILanguageRuntimeMessageOutput, ILanguageRuntimeMessagePrompt, ILanguageRuntimeMessageState, ILanguageRuntimeMessageStream, ILanguageRuntimeMetadata, ILanguageRuntimeService, RuntimeCodeExecutionMode, RuntimeErrorBehavior, RuntimeOnlineState, RuntimeState } from '../../../services/languageRuntime/common/languageRuntimeService.js';
-import { ILanguageRuntimeSession, IRuntimeSessionService } from '../../../services/runtimeSession/common/runtimeSessionService.js';
+import { ILanguageRuntimeMetadata, ILanguageRuntimeService } from '../../../services/languageRuntime/common/languageRuntimeService.js';
+import { IRuntimeSessionService } from '../../../services/runtimeSession/common/runtimeSessionService.js';
 import { IRuntimeStartupService } from '../../../services/runtimeStartup/common/runtimeStartupService.js';
-import { NotebookCellTextModel } from '../../notebook/common/model/notebookCellTextModel.js';
+import { IPYNB_VIEW_TYPE } from '../../notebook/browser/notebookBrowser.js';
 import { NotebookTextModel } from '../../notebook/common/model/notebookTextModel.js';
-import { IOutputItemDto } from '../../notebook/common/notebookCommon.js';
-import { CellExecutionUpdateType } from '../../notebook/common/notebookExecutionService.js';
-import { INotebookCellExecution, INotebookExecutionStateService } from '../../notebook/common/notebookExecutionStateService.js';
-import { INotebookKernel, INotebookKernelChangeEvent, INotebookKernelService, VariablesResult } from '../../notebook/common/notebookKernelService.js';
+import { INotebookKernelService } from '../../notebook/common/notebookKernelService.js';
 import { INotebookService } from '../../notebook/common/notebookService.js';
+import { ActiveNotebookHasRunningRuntimeManager } from '../common/activeNotebookHasRunningRuntime.js';
 import { registerRuntimeNotebookKernelActions } from '../common/runtimeNotebookKernelActions.js';
-import { isRuntimeNotebookKernelEnabled, POSITRON_RUNTIME_NOTEBOOK_KERNELS_EXTENSION_ID } from '../common/runtimeNotebookKernelConfig.js';
+import { isRuntimeNotebookKernelEnabled } from '../common/runtimeNotebookKernelConfig.js';
 import { IRuntimeNotebookKernelService } from './interfaces/runtimeNotebookKernelService.js';
 import { NotebookExecutionStatus } from './notebookExecutionStatus.js';
-import { ActiveNotebookHasRunningRuntimeManager } from '../common/activeNotebookHasRunningRuntime.js';
-
-/**
- * The view type supported by Positron runtime notebook kernels. Currently only Jupyter notebooks are supported.
- */
-const viewType = 'jupyter-notebook';
+import { RuntimeNotebookKernel } from './runtimeNotebookKernel.js';
 
 /**
  * The affinity of a kernel for a notebook.
@@ -49,28 +33,6 @@ enum NotebookKernelAffinity {
 
 	/** A kernel will be automatically started if it is a notebook's only preferred kernel. */
 	Preferred = 2
-}
-
-/**
- * The type of a Jupyter notebook cell output.
- *
- * Used by the ipynb notebook serializer (extensions/ipynb/src/serializers.ts) to convert from
- * VSCode notebook cell outputs to Jupyter notebook cell outputs.
- *
- * See: https://jupyter-client.readthedocs.io/en/latest/messaging.html
- */
-enum JupyterNotebookCellOutputType {
-	/** An error occurred during an execution. */
-	Error = 'error',
-
-	/** Output from one of the standard streams (stdout or stderr). */
-	Stream = 'stream',
-
-	/** One of possibly many outputs related to an execution. */
-	DisplayData = 'display_data',
-
-	/** The result of an execution. */
-	ExecuteResult = 'execute_result',
 }
 
 export class RuntimeNotebookKernelService extends Disposable implements IRuntimeNotebookKernelService {
@@ -155,8 +117,8 @@ export class RuntimeNotebookKernelService extends Disposable implements IRuntime
 
 		// Register kernel source action providers. This is how we customize the
 		// kernel selection quickpick. Each command must return a valid runtime ID.
-		this._register(this._notebookKernelService.registerKernelSourceActionProvider(viewType, {
-			viewType,
+		this._register(this._notebookKernelService.registerKernelSourceActionProvider(IPYNB_VIEW_TYPE, {
+			viewType: IPYNB_VIEW_TYPE,
 			async provideKernelSourceActions() {
 				return [
 					{
@@ -266,483 +228,6 @@ export class RuntimeNotebookKernelService extends Disposable implements IRuntime
 	 * Placeholder that gets called to "initialize" the service.
 	 */
 	initialize(): void {
-	}
-}
-
-// TODO: Move each class to a separate file.
-export class RuntimeNotebookKernel extends Disposable implements INotebookKernel {
-	public readonly viewType = viewType;
-
-	public readonly extension = new ExtensionIdentifier(POSITRON_RUNTIME_NOTEBOOK_KERNELS_EXTENSION_ID);
-
-	public readonly preloadUris: URI[] = [];
-
-	public readonly preloadProvides: string[] = [];
-
-	public readonly implementsInterrupt = true;
-
-	public readonly implementsExecutionOrder = true;
-
-	public readonly hasVariableProvider = false;
-
-	// TODO: Not sure what we could set this to...
-	public readonly localResourceRoot: URI = URI.parse('');
-
-	private readonly _onDidChange = this._register(new Emitter<INotebookKernelChangeEvent>());
-
-	/** An event that fires when the kernel's details change. */
-	public readonly onDidChange = this._onDidChange.event;
-
-	private readonly _notebookRuntimeKernelSessionsByNotebookUri = new ResourceMap<RuntimeNotebookKernelSession>();
-
-	constructor(
-		public readonly runtime: ILanguageRuntimeMetadata,
-		@IInstantiationService private readonly _instantiationService: IInstantiationService,
-		@ILogService private readonly _logService: ILogService,
-		@INotebookService private readonly _notebookService: INotebookService,
-		@IRuntimeSessionService private readonly _runtimeSessionService: IRuntimeSessionService,
-	) {
-		super();
-	}
-
-	get id(): string {
-		// This kernel ID format is assumed by a few services and should be changed carefully.
-		return `${this.extension.value}/${this.runtime.runtimeId}`;
-	}
-
-	get label(): string {
-		return this.runtime.runtimeName;
-	}
-
-	get description(): string {
-		return this.runtime.runtimePath;
-	}
-
-	get detail(): string | undefined {
-		return undefined;
-	}
-
-	get supportedLanguages(): string[] {
-		return [this.runtime.languageId, 'raw'];
-	}
-
-	async executeNotebookCellsRequest(notebookUri: URI, cellHandles: number[]): Promise<void> {
-		this._logService.debug(`[RuntimeNotebookKernel] Executing cells: ${cellHandles.join(', ')}`);
-
-		const notebook = this._notebookService.getNotebookTextModel(notebookUri);
-		if (!notebook) {
-			// Copying ExtHostNotebookController.getNotebookDocument for now.
-			throw new Error(`NO notebook document for '${notebookUri}'`);
-		}
-
-		const session = this._runtimeSessionService.getNotebookSessionForNotebookUri(notebookUri);
-		if (!session) {
-			throw new Error(`NO runtime session for notebook '${notebookUri}'`);
-		}
-
-		// TODO: Dispose?
-		let notebookRuntimeKernelSession = this._notebookRuntimeKernelSessionsByNotebookUri.get(notebookUri);
-		if (!notebookRuntimeKernelSession) {
-			notebookRuntimeKernelSession = this._instantiationService.createInstance(RuntimeNotebookKernelSession, session, notebook);
-			this._notebookRuntimeKernelSessionsByNotebookUri.set(notebookUri, notebookRuntimeKernelSession);
-		}
-
-		// Get the cells to execute from their handles.
-		const cells: NotebookCellTextModel[] = [];
-		for (const cellHandle of cellHandles) {
-			const cell = notebook.cells.find(cell => cell.handle === cellHandle);
-			// TODO: When does this happen?
-			if (!cell) {
-				continue;
-			}
-			cells.push(cell);
-		}
-
-		// Execute the cells.
-		try {
-			await notebookRuntimeKernelSession.executeCells(cells);
-		} catch (err) {
-			this._logService.debug(`Error executing cells: ${err.stack ?? err}`);
-		}
-	}
-
-	async cancelNotebookCellExecution(uri: URI, _cellHandles: number[]): Promise<void> {
-		this._logService.debug(`[RuntimeNotebookKernel] Interrupting`);
-
-		const notebookRuntimeKernelSession = this._notebookRuntimeKernelSessionsByNotebookUri.get(uri);
-		if (notebookRuntimeKernelSession) {
-			notebookRuntimeKernelSession.interrupt();
-		}
-	}
-
-	provideVariables(notebookUri: URI, parentId: number | undefined, kind: 'named' | 'indexed', start: number, token: CancellationToken): AsyncIterableObject<VariablesResult> {
-		throw new Error('provideVariables not implemented.');
-	}
-
-	public override dispose(): void {
-		super.dispose();
-
-		for (const disposable of this._notebookRuntimeKernelSessionsByNotebookUri.values()) {
-			disposable.dispose();
-		}
-	}
-}
-
-class RuntimeNotebookKernelSession extends Disposable {
-	/**
-	 * A map of the last queued cell execution promise for each notebook, keyed by notebook URI.
-	 * Each queued cell execution promise is chained to the previous one for the notebook,
-	 * so that cells are executed in order.
-	 */
-	private readonly _pendingCellExecutionsByNotebookUri = new ResourceMap<Promise<void>>();
-
-	private _pendingRuntimeExecution: RuntimeNotebookCellExecution | undefined;
-
-	constructor(
-		private readonly _session: ILanguageRuntimeSession,
-		private readonly _notebook: NotebookTextModel,
-		@IInstantiationService private readonly _instantiationService: IInstantiationService,
-		@ILogService private readonly _logService: ILogService,
-		@INotebookExecutionStateService private readonly _notebookExecutionStateService: INotebookExecutionStateService,
-	) {
-		super();
-	}
-
-	async executeCells(cells: NotebookCellTextModel[]): Promise<void> {
-		const executionPromises = cells.map(cell => this.queueCellExecution(cell));
-		await Promise.all(executionPromises);
-	}
-
-	private async queueCellExecution(cell: NotebookCellTextModel): Promise<void> {
-		// Get the pending execution for this notebook, if one exists.
-		const pendingExecution = this._pendingCellExecutionsByNotebookUri.get(this._notebook.uri);
-
-		// Chain this execution after the pending one.
-		const currentExecution = Promise.resolve(pendingExecution)
-			.then(() => this.executeCell(cell))
-			.finally(() => {
-				// If this was the last execution in the chain, remove it from the map,
-				// starting a new chain.
-				if (this._pendingCellExecutionsByNotebookUri.get(this._notebook.uri) === currentExecution) {
-					this._pendingCellExecutionsByNotebookUri.delete(this._notebook.uri);
-				}
-			});
-
-		// Update the pending execution for this notebook.
-		this._pendingCellExecutionsByNotebookUri.set(this._notebook.uri, currentExecution);
-
-		return currentExecution;
-	}
-
-	private async executeCell(cell: NotebookCellTextModel): Promise<void> {
-		// Don't try to execute raw cells; they're often used to define metadata e.g in Quarto notebooks.
-		if (cell.language === 'raw') {
-			return;
-		}
-
-		const code = cell.getValue();
-
-		// If the cell is empty, skip it.
-		if (!code.trim()) {
-			return;
-		}
-
-		const cellExecution = this._notebookExecutionStateService.getCellExecution(cell.uri);
-		if (!cellExecution) {
-			throw new Error(`NO execution for cell '${cell.uri}'`);
-		}
-
-		// TODO: This can be a simple counter for the session.
-		const id = generateUuid();
-
-		// TODO: Error if there's already a runtimeExecution for this ID?
-		//       Or get another id?
-
-		const execution = this._register(this._instantiationService.createInstance(
-			RuntimeNotebookCellExecution, this._session, cellExecution, cell
-		));
-		this._pendingRuntimeExecution = execution;
-
-		try {
-			this._session.execute(
-				code,
-				id,
-				RuntimeCodeExecutionMode.Interactive,
-				RuntimeErrorBehavior.Stop,
-			);
-		} catch (err) {
-			execution.error(err);
-		}
-
-		return this._pendingRuntimeExecution.promise.finally(() => {
-			if (this._pendingRuntimeExecution === execution) {
-				this._pendingRuntimeExecution = undefined;
-			}
-		});
-	}
-
-	async interrupt(): Promise<void> {
-		if (this._session.getRuntimeState() === RuntimeState.Busy ||
-			this._session.getRuntimeState() === RuntimeState.Interrupting) {
-			// The session is in an interruptible state, interrupt it.
-			this._session.interrupt();
-		} else if (this._session.getRuntimeState() === RuntimeState.Exiting ||
-			this._session.getRuntimeState() === RuntimeState.Exited ||
-			this._session.getRuntimeState() === RuntimeState.Restarting ||
-			this._session.getRuntimeState() === RuntimeState.Uninitialized) {
-
-			// TODO: Is it possible that a user could interrupt without a RuntimeNotebookKernelSession at all?
-			//       That would leave the executions running atm.
-
-			const execution = this._pendingRuntimeExecution;
-			if (!execution) {
-				// It shouldn't be possible to interrupt an execution without having set a token source.
-				// Log a warning and do nothing.
-				this._logService.warn(`Tried to interrupt notebook ${this._notebook.uri.toString()} with no executing cell.`);
-				return;
-			}
-
-			// It's possible that the session exited after the execution started.
-			// Log a warning and cancel the execution promise.
-			// TODO: Should this be primarily handled in a session.onDidEndSession listener?
-			this._logService.warn(`Tried to interrupt notebook ${this._notebook.uri.toString()} with no running session. Cancelling execution.`);
-			execution.error({
-				name: 'Session Exited Unexpectedly',
-				message: 'The session has exited unexpectedly.',
-			});
-		}
-	}
-}
-
-class RuntimeNotebookCellExecution extends Disposable {
-	private _deferred = new DeferredPromise<void>();
-
-	// Create a promise tracking the current message for the cell. Each execution may
-	// receive multiple messages, which we want to handle in sequence.
-	private _taskQueue = new Sequencer();
-
-	constructor(
-		private readonly _session: ILanguageRuntimeSession,
-		private readonly _cellExecution: INotebookCellExecution,
-		private readonly _cell: NotebookCellTextModel,
-		@ILogService private readonly _logService: ILogService,
-		@IQuickInputService private readonly _quickInputService: IQuickInputService,
-	) {
-		super();
-
-		this._register(this._session.onDidReceiveRuntimeMessageInput(async message => {
-			await this._taskQueue.queue(() => this.handleRuntimeMessageInput(message));
-		}));
-
-		this._register(this._session.onDidReceiveRuntimeMessagePrompt(async message => {
-			await this._taskQueue.queue(() => this.handleRuntimeMessagePrompt(message));
-		}));
-
-		this._register(this._session.onDidReceiveRuntimeMessageOutput(async message => {
-			await this._taskQueue.queue(() => this.handleRuntimeMessageOutput(
-				message, JupyterNotebookCellOutputType.DisplayData
-			));
-		}));
-
-		this._register(this._session.onDidReceiveRuntimeMessageResult(async message => {
-			await this._taskQueue.queue(() => this.handleRuntimeMessageOutput(
-				message, JupyterNotebookCellOutputType.ExecuteResult
-			));
-		}));
-
-		this._register(this._session.onDidReceiveRuntimeMessageStream(async message => {
-			await this._taskQueue.queue(() => this.handleRuntimeMessageStream(message));
-		}));
-
-		this._register(this._session.onDidReceiveRuntimeMessageError(async message => {
-			await this._taskQueue.queue(() => this.handleRuntimeMessageError(message));
-		}));
-
-		this._register(this._session.onDidReceiveRuntimeMessageState(async message => {
-			await this._taskQueue.queue(() => this.handleRuntimeMessageState(message));
-		}));
-
-		this._cellExecution.update([{
-			// Start the execution timer.
-			editType: CellExecutionUpdateType.ExecutionState,
-			runStartTime: Date.now(),
-		}, {
-			// Clear any existing outputs.
-			editType: CellExecutionUpdateType.Output,
-			cellHandle: this._cellExecution.cellHandle,
-			outputs: [],
-		}]);
-	}
-
-	public complete(): void {
-		// End the cell execution.
-		this._cellExecution.complete({
-			runEndTime: Date.now(),
-			lastRunSuccess: true,
-		});
-
-		this._deferred.complete();
-
-		// TODO: Do we need to track whether we're disposed in some cases?
-		this.dispose();
-	}
-
-	public error(err: Error): void {
-		// End the cell execution with the error.
-		this._cellExecution.complete({
-			runEndTime: Date.now(),
-			lastRunSuccess: false,
-			error: {
-				message: err.message,
-				stack: err?.stack ?? JSON.stringify(err),
-				uri: this._cell.uri,
-				location: undefined,
-			},
-		});
-
-		this._deferred.error(err);
-
-		// TODO: Do we need to track whether we're disposed in some cases?
-		this.dispose();
-	}
-
-	private async handleRuntimeMessageInput(message: ILanguageRuntimeMessageInput): Promise<void> {
-		// Update the cell's execution order (usually displayed in notebook UIs).
-		this._cellExecution.update([{
-			editType: CellExecutionUpdateType.ExecutionState,
-			executionOrder: message.execution_count,
-		}]);
-	}
-
-	private async handleRuntimeMessagePrompt(message: ILanguageRuntimeMessagePrompt): Promise<void> {
-		// Let the user input a reply.
-		const reply = await this._quickInputService.input({
-			password: message.password,
-			prompt: message.prompt,
-		});
-
-		// Reply to the prompt.
-		this._session.replyToPrompt(message.id, reply ?? '');
-	}
-
-	private async handleRuntimeMessageOutput(
-		message: ILanguageRuntimeMessageOutput,
-		outputType: JupyterNotebookCellOutputType,
-	): Promise<void> {
-		const outputItems: IOutputItemDto[] = [];
-		for (const [mime, data] of Object.entries(message.data)) {
-			switch (mime) {
-				case 'image/png':
-				case 'image/jpeg':
-					outputItems.push({ data: decodeBase64(String(data)), mime });
-					break;
-				// This list is a subset of src/vs/workbench/contrib/notebook/browser/view/cellParts/cellOutput.JUPYTER_RENDERER_MIMETYPES
-				case 'application/geo+json':
-				case 'application/vdom.v1+json':
-				case 'application/vnd.dataresource+json':
-				case 'application/vnd.jupyter.widget-view+json':
-				case 'application/vnd.plotly.v1+json':
-				case 'application/vnd.r.htmlwidget':
-				case 'application/vnd.vega.v2+json':
-				case 'application/vnd.vega.v3+json':
-				case 'application/vnd.vega.v4+json':
-				case 'application/vnd.vega.v5+json':
-				case 'application/vnd.vegalite.v1+json':
-				case 'application/vnd.vegalite.v2+json':
-				case 'application/vnd.vegalite.v3+json':
-				case 'application/vnd.vegalite.v4+json':
-				case 'application/x-nteract-model-debug+json':
-					// The JSON cell output item will be rendered using the appropriate notebook renderer.
-					outputItems.push({ data: VSBuffer.fromString(JSON.stringify(data, undefined, '\t')), mime });
-					break;
-				default:
-					outputItems.push({ data: VSBuffer.fromString(String(data)), mime });
-			}
-		}
-		this._cellExecution.update([{
-			editType: CellExecutionUpdateType.Output,
-			cellHandle: this._cellExecution.cellHandle,
-			append: true,
-			outputs: [{
-				outputId: message.id,
-				outputs: outputItems,
-				metadata: { outputType },
-			}]
-		}]);
-	}
-
-	private async handleRuntimeMessageStream(message: ILanguageRuntimeMessageStream): Promise<void> {
-		// Convert the runtime message into an output item data transfer object.
-		let mime: string;
-		if (message.name === 'stdout') {
-			mime = 'application/vnd.code.notebook.stdout';
-		} else if (message.name === 'stderr') {
-			mime = 'application/vnd.code.notebook.stderr';
-		} else {
-			this._logService.warn(`[NotebookRuntimeKernel] Ignoring runtime message with unknown stream name: ${message.name}`);
-			return;
-		}
-		const newOutputItem: IOutputItemDto = { data: VSBuffer.fromString(message.text), mime };
-
-		// If the last output has items of the same mime type (i.e. from the same stream: stdout/stderr),
-		// append the new item to the last output. Otherwise, create a new output.
-		const lastOutput = this._cell.outputs.at(-1);
-		const lastOutputItems = lastOutput?.outputs;
-		if (lastOutputItems && lastOutputItems.every(item => item.mime === mime)) {
-			this._cellExecution.update([{
-				editType: CellExecutionUpdateType.OutputItems,
-				append: true,
-				outputId: lastOutput.outputId,
-				items: [newOutputItem],
-			}]);
-		} else {
-			this._cellExecution.update([{
-				editType: CellExecutionUpdateType.Output,
-				cellHandle: this._cellExecution.cellHandle,
-				append: true,
-				outputs: [{
-					outputId: message.id,
-					outputs: [newOutputItem],
-					metadata: { outputType: JupyterNotebookCellOutputType.Stream },
-				}]
-			}]);
-		}
-	}
-
-	private async handleRuntimeMessageError(message: ILanguageRuntimeMessageError): Promise<void> {
-		this._cellExecution.update([{
-			editType: CellExecutionUpdateType.Output,
-			cellHandle: this._cellExecution.cellHandle,
-			append: true,
-			outputs: [{
-				outputId: message.id,
-				outputs: [{
-					data: VSBuffer.fromString(JSON.stringify({
-						name: message.name,
-						message: message.message,
-						stack: message.traceback.join('\n'),
-					}, undefined, '\t')),
-					mime: 'application/vnd.code.notebook.error',
-				}],
-				metadata: { outputType: JupyterNotebookCellOutputType.Error },
-			}],
-		}]);
-
-		this.error({
-			name: message.name,
-			message: message.message,
-			stack: message.traceback.join('\n'),
-		});
-	}
-
-	private async handleRuntimeMessageState(message: ILanguageRuntimeMessageState): Promise<void> {
-		if (message.state === RuntimeOnlineState.Idle) {
-			this.complete();
-		}
-	}
-
-	public get promise(): Promise<void> {
-		return this._deferred.p;
 	}
 }
